requirements.txt
flask
torch
shap-e @ git+https://github.com/openai/shap-e.git
trimesh
pygltflib
numpy
pillow
pyyaml
uuid
gunicorn